feat: meeting MemberInfoResponse 변경

+++ backend/booking/src/main/java/com/booking/booking/post/domain/Post.java
@@ -30,7 +30,7 @@ public class Post {
     @JoinColumn(name = "meetingId")
     private Meeting meeting;
 
-    private long memberId;
+    private Integer memberId;
 
     private String title;
